pyproject.toml
[project]
name = "dynamicformagent"
version = "0.1.0"
description = "A multi-agent system to dynamically fill a JSON object through natural conversation."
authors = [
    {name = "Anass El Achham",email = "anas.elachham@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph (>=0.3.2,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "together (>=1.4.1,<2.0.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "langchain-core (>=0.3.39,<0.4.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "langgraph-prebuilt (>=0.1.1,<0.2.0)",
    "langgraph-swarm (>=0.0.4,<0.0.5)",
    "python-fasthtml (>=0.12.4,<0.13.0)",
    "flask (>=3.1.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

